character.rbファイルの作成
